% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clean_dfp_itr_data.R
\name{clean_dfp_itr_data}
\alias{clean_dfp_itr_data}
\title{Cleans ITR/DFP table}
\usage{
clean_dfp_itr_data(df_in, file_in)
}
\arguments{
\item{df_in}{A itr/dpf dataframe}

\item{file_in}{path of csv file}
}
\value{
A Clean dataframe
}
\description{
Cleans ITR/DFP table
}
\examples{

\dontrun{
# no example
}

}
