% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download_read_dfp_zip_file.R
\name{download_read_dfp_zip_file}
\alias{download_read_dfp_zip_file}
\title{Downloads and reads DFP datasets}
\usage{
download_read_dfp_zip_file(
  url_in,
  companies_cvm_codes,
  type_format,
  cache_folder = "gcvmd_cache",
  clean_data
)
}
\arguments{
\item{url_in}{Url of file to download and read}

\item{companies_cvm_codes}{Numeric CVM code  of companies. IF set to NULL (default), will return data for all available companies.}

\item{type_format}{Type of format of document (con = consolidated, ind = individual). Default = c('con', 'ind')}

\item{cache_folder}{Path of cache folder to keep memoise and zip files}

\item{clean_data}{Clean data or return raw data? See read_dfp|itr_csv() for details}
}
\value{
A dataframe with dfp data
}
\description{
Downloads and reads DFP datasets
}
\examples{

\dontrun{ # keep cran check fast
my_url <- 'http://dados.cvm.gov.br/dados/CIA_ABERTA/DOC/DFP/DRE/DADOS/dre_cia_aberta_2010.zip'
df_out <- download_read_dfp_zip_file(url_in = my_url,
                                     companies_cvm_codes = NULL,
                                     type_format = 'con',
                                     clean_data = TRUE)
}
}
