% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_export_xlsx.R
\name{export_xlsx}
\alias{export_xlsx}
\title{Export DFP/ITR data to a xlsx file}
\usage{
export_xlsx(l_dfp, f_xlsx = "GetDFPData-XLSX.xlsx")
}
\arguments{
\item{l_dfp}{A list from get_dfp_data or get_itr_data}

\item{f_xlsx}{Path to xlsx file}
}
\value{
A dataframe with several information about B3 companies
}
\description{
Export DFP/ITR data to a xlsx file
}
\examples{

\dontrun{ # keep cran check fast
df_info <- get_info_companies()
str(df_info)
}
}
