% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download_file.R
\name{find_dl_size}
\alias{find_dl_size}
\title{Find downloas size}
\usage{
find_dl_size(url_in)
}
\arguments{
\item{url_in}{An internet address}
}
\value{
size of download file
}
\description{
Find downloas size
}
\examples{
find_dl_size('https://www.msperlin.com/blog/files/adfer-files/TOC-adfeR-ed3.pdf')

}
