% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{fix_cnpj}
\alias{fix_cnpj}
\title{Converts cnpj from text to numeric}
\usage{
fix_cnpj(cnpj_in)
}
\arguments{
\item{cnpj_in}{The character (single) of cnpj}
}
\value{
A numeric cnpj
}
\description{
Converts cnpj from text to numeric
}
\examples{

fix_cnpj('00.111.004/005')

}
