% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_ftp_contents.R
\name{get_contents_ftp}
\alias{get_contents_ftp}
\title{Fetches ftp contents}
\usage{
get_contents_ftp(ftp_url)
}
\arguments{
\item{ftp_url}{The url of ftp}
}
\value{
A dataframe with links and files
}
\description{
Fetches ftp contents
}
\examples{
\dontrun{ # keep cran fast
df_ftp <- get_contents_ftp('http://dados.cvm.gov.br/dados/CIA_ABERTA/DOC/DFP/DRE/DADOS/')
}
}
