% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_dfp_docs.R
\name{get_dfp_docs}
\alias{get_dfp_docs}
\title{Get data from dfp docs}
\usage{
get_dfp_docs(
  companies_cvm_codes,
  type_docs,
  type_format,
  first_year,
  last_year,
  clean_data,
  use_memoise,
  cache_folder
)
}
\arguments{
\item{companies_cvm_codes}{Numeric CVM code  of companies. IF set to NULL (default), will return data for all available companies.}

\item{type_docs}{Type of financial documents. E.g. c('DRE', 'BPA'). Definitions: '*' = fetch all docs,  'BPA' = Assets (ativos),
'BPP' = Liabilities (passivo),
'DRE' = income statement (demonstrativo de resultados),
'DFC_MD' = cash flow by direct method (fluxo de caixa pelo metodo direto),
'DFC_MI' = cash flow by indirect method (fluxo de caixa pelo metodo indireto),
'DMPL' = statement of changes in equity (mutacoes do patrimonio liquido),
'DVA' = value added report (desmonstrativo de valor agregado)}

\item{type_format}{Type of format of document (con = consolidated, ind = individual). Default = c('con', 'ind')}

\item{first_year}{First year of selected data}

\item{last_year}{Last year of selected data}

\item{clean_data}{Clean data or return raw data? See read_dfp|itr_csv() for details}

\item{use_memoise}{Use memoise caching? If no (default), the function will read all .csv files. If yes, will use package
memoise for caching results (execution speed increases significantly)}

\item{cache_folder}{Path of cache folder to keep memoise and zip files}
}
\value{
A dataframe
}
\description{
Get data from dfp docs
}
\examples{
\dontrun{
# no example
}
}
