% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_tickers.R
\name{get_tickers}
\alias{get_tickers}
\title{Retrieves a ticker list from cvm open data repository}
\usage{
get_tickers(year = lubridate::year(Sys.Date()))
}
\arguments{
\item{year}{year of the data (2018 onwards)}
}
\value{
a dataframe with information about companies and tickers
}
\description{
Retrieves ticker information from <https://dados.cvm.gov.br/dados/CIA_ABERTA/DOC/FCA/DADOS/>.
}
\examples{
\dontrun{
tickers <- get_tickers()
}
}
