% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_dfp_csv.R
\name{read_dfp_csv}
\alias{read_dfp_csv}
\title{Reads DFP csv file}
\usage{
read_dfp_csv(file_in, clean_data)
}
\arguments{
\item{file_in}{path of csv file}

\item{clean_data}{Clean data or return raw data? See read_dfp|itr_csv() for details}
}
\value{
A dataframe
}
\description{
Reads DFP csv file
}
\examples{

\dontrun{
# no example
}

}
