% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_search_company.R
\name{search_company}
\alias{search_company}
\title{Helps users search for a company}
\usage{
search_company(char_to_search, cache_folder = "gcvmd_cache")
}
\arguments{
\item{char_to_search}{Character for partial matching}

\item{cache_folder}{Path of cache folder to keep memoise and zip files}
}
\value{
A dataframe with companies
}
\description{
Helps users search for a company
}
\examples{

\dontrun{ # dontrun: keep cran check fast
df <- search_company('petrobras')
}
}
