% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{select_responsible_beverage}
\alias{select_responsible_beverage}
\title{Makes sure people dont drink at afternoon}
\usage{
select_responsible_beverage()
}
\value{
A responsible drink
}
\description{
Makes sure people dont drink at afternoon
}
\examples{
print(select_responsible_beverage())
}
