% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_info_companies.R
\name{get_info_companies}
\alias{get_info_companies}
\title{Returns a up to date table with information about B3 companies}
\usage{
get_info_companies(cache_folder = "gcvmd_cache")
}
\arguments{
\item{cache_folder}{Path of cache folder to keep memoise and zip files}
}
\value{
A dataframe with several information about B3 companies
}
\description{
Imports a dataset from cvm <http://sistemas.cvm.gov.br/cadastro/SPW_CIA_ABERTA.ZIP>, containing
up to date information about companies, active or not.
}
\examples{

\dontrun{ # keep cran check fast
df_info <- get_info_companies()
str(df_info)
}
}
